[FIX] sale,product_email_template: match the help with the behaviour
Before this commit, the help of the product_email_template module said
that the email is sent when the invoice is paid. This is not the case,
the current behaviour of the module is that the email is sent when the
invoice is validated.

Fine-tuning of : 976bd310bfd05e24cb78b0e669405781182e6168,
7ca768777b5fc504c3e7a1e06818ba3c901ed946
and 330d0167f7007622828a4403a3dfdc56d0053aa5

opw-1973946

closes odoo/odoo#33686

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>